+++ packages/common/src/invitationCode.ts
@@ -1,8 +1,6 @@
 import { InvitationParams, Site } from './static'
 
-type PotentialString = string | undefined | null
-
-export const retrieveInvitationCode = (url: string): PotentialString => {
+export const retrieveInvitationCode = (url: string): string => {
   /**
    * Extract invitation code from deep url.
    * Valid format: quiet://?code=<invitation code>
@@ -11,20 +9,22 @@ export const retrieveInvitationCode = (url: string): PotentialString => {
   try {
     data = new URL(url)
   } catch (e) {
-    return
+    return ''
   }
-  if (!data || data.protocol !== 'quiet:') return null
-  if (data.searchParams.has(InvitationParams.CODE)) {
-    console.log('Retrieved code:', data.searchParams.get(InvitationParams.CODE))
-    return data.searchParams.get(InvitationParams.CODE)
+  if (!data || data.protocol !== 'quiet:') return ''
+  const code = data.searchParams.get(InvitationParams.CODE)
+  if (code) {
+    console.log('Retrieved code:', code)
+    return code
   }
+  return ''
 }
 
-export const argvInvitationCode = (argv: string[]): PotentialString => {
+export const argvInvitationCode = (argv: string[]): string => {
   /**
    * Extract invitation code from deep url if url is present in argv
    */
-  let invitationCode: PotentialString
+  let invitationCode: string = ''
   for (const arg of argv) {
     invitationCode = retrieveInvitationCode(arg)
     if (invitationCode) {