+++ packages/desktop/src/renderer/components/Channel/File/UploadedImage/UploadedImage.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react'
 import { styled } from '@mui/material/styles'
-import { DisplayableMessage, DownloadStatus } from '@quiet/state-manager'
-import { UseModalTypeWrapper } from '../../../../containers/hooks'
+import { DownloadStatus, FileMetadata } from '@quiet/types'
+import { UseModalType } from '../../../../containers/hooks'
 import UploadedFileModal from './UploadedImagePreview'
 import { UploadedFilename, UploadedImagePlaceholder } from '../UploadedImagePlaceholder/UploadedImagePlaceholder'
 
@@ -25,24 +25,21 @@ const Root = styled('div')(() => ({
 }))
 
 export interface UploadedImageProps {
-  message: DisplayableMessage
-  uploadedFileModal?: ReturnType<
-  UseModalTypeWrapper<{
-    src: string
-  }>['types']
-  >
-  downloadStatus: DownloadStatus
+  media: FileMetadata
+  uploadedFileModal?:
+    UseModalType<{
+      src: string
+    }>
+
+  downloadStatus?: DownloadStatus
 }
 
-export const UploadedImage: React.FC<UploadedImageProps> = ({ message, uploadedFileModal, downloadStatus }) => {
+export const UploadedImage: React.FC<UploadedImageProps> = ({ media, uploadedFileModal, downloadStatus }) => {
   const [showImage, setShowImage] = useState<boolean>(false)
+  const { cid, path, name, ext } = media
 
-  const { cid, path, name, ext } = message.media
-
-  const imageWidth = message.media?.width
-  const imageHeight = message.media?.height
-
-  const width = imageWidth >= 400 ? 400 : imageWidth
+  const imageWidth = media.width
+  const imageHeight = media.height
 
   useEffect(() => {
     if (uploadedFileModal?.open) {
@@ -51,12 +48,15 @@ export const UploadedImage: React.FC<UploadedImageProps> = ({ message, uploadedF
   }, [uploadedFileModal?.open])
 
   useEffect(() => {
-    if (showImage) {
+    if (showImage && path) {
       uploadedFileModal?.handleOpen({
         src: path
       })
     }
   }, [showImage])
+
+  if (!imageWidth || !imageHeight) return null
+  const width = imageWidth >= 400 ? 400 : imageWidth
 
   return (
     (<Root>
@@ -76,7 +76,7 @@ export const UploadedImage: React.FC<UploadedImageProps> = ({ message, uploadedF
               />
             </div>
           </div>
-          <UploadedFileModal {...uploadedFileModal} uploadedFileModal={uploadedFileModal} />
+          {uploadedFileModal && <UploadedFileModal {...uploadedFileModal} uploadedFileModal={uploadedFileModal} />}
         </>
       ) : (
         <UploadedImagePlaceholder