+++ packages/desktop/src/renderer/components/widgets/channels/BasicMessage.tsx
@@ -12,16 +12,16 @@ import red from '@mui/material/colors/red'
 
 import Jdenticon from '../../Jdenticon/Jdenticon'
 
-import { DisplayableMessage, DownloadStatus, MessageSendingStatus } from '@quiet/state-manager'
+import { DisplayableMessage, DownloadStatus, MessageSendingStatus } from '@quiet/types'
 
 import { NestedMessageContent } from './NestedMessageContent'
-import { UseModalTypeWrapper } from '../../../containers/hooks'
 
 import { FileActionsProps } from '../../Channel/File/FileComponent/FileComponent'
 
 import information from '../../../static/images/updateIcon.svg'
 
 import Icon from '../../ui/Icon/Icon'
+import { UseModalType } from '../../../containers/hooks'
 
 const PREFIX = 'BasicMessageComponent'
 
@@ -152,11 +152,9 @@ export interface BasicMessageProps {
   pendingMessages?: Dictionary<MessageSendingStatus>
   openUrl: (url: string) => void
   downloadStatuses?: Dictionary<DownloadStatus>
-  uploadedFileModal?: ReturnType<
-  UseModalTypeWrapper<{
+  uploadedFileModal?: UseModalType<{
     src: string
-  }>['types']
-  >
+  }>
   onMathMessageRendered?: () => void
 }