+++ packages/desktop/src/rtl-tests/channel.menu.test.tsx
@@ -40,7 +40,7 @@ describe('Channel menu', () => {
     await factory.create<ReturnType<typeof communities.actions.addNewCommunity>['payload']>(
       'Community',
       {
-        id: 0,
+        id: '0',
         name: 'community',
         CA: null,
         registrarUrl: 'http://ugmx77q2tnm5fliyfxfeen5hsuzjtbsz44tsldui2ju7vl5xj4d447yd.onion',