+++ packages/mobile/src/components/Message/Message.types.ts
@@ -1,10 +1,10 @@
-import { DisplayableMessage, DownloadStatus, FileMetadata, MessageSendingStatus } from '@quiet/state-manager'
+import { DisplayableMessage, DownloadStatus, FileMetadata, MessageSendingStatus } from '@quiet/types'
 import { Dictionary } from '@reduxjs/toolkit'
 
 export interface MessageProps {
   data: DisplayableMessage[]
   pendingMessages?: Dictionary<MessageSendingStatus>
   downloadStatus?: DownloadStatus
-  openImagePreview?: (medi: FileMetadata) => void
-  openUrl?: (url: string) => void
+  openImagePreview: (media: FileMetadata) => void
+  openUrl: (url: string) => void
 }