+++ packages/state-manager/src/sagas/files/files.types.ts
@@ -55,6 +55,7 @@ export interface DownloadProgress {
 }
 
 export enum DownloadState {
+  None = '',
   Uploading = 'uploading',
   Hosted = 'hosted',
   Ready = 'ready',