+++ packages/state-manager/src/sagas/publicChannels/publicChannels.slice.ts
@@ -53,6 +53,7 @@ export const publicChannelsSlice = createSlice({
   reducers: {
     createChannel: (state, _action: PayloadAction<CreateChannelPayload>) => state,
     deleteChannel: (state, _action: PayloadAction<DeleteChannelPayload>) => state,
+    completeChannelDeletion: (state, _action) => state,
     channelDeletionResponse: (state, _action: PayloadAction<ChannelDeletionResponsePayload>) =>
       state,
     deleteChannelFromStore: (state, action: PayloadAction<DeleteChannelFromStorePayload>) => {