+++ packages/common/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.1-alpha.0](https://github.com/TryQuiet/quiet/compare/@quiet/common@1.3.0...@quiet/common@1.3.1-alpha.0) (2023-06-06)
+
+**Note:** Version bump only for package @quiet/common
+
+
+
+
+
+# [1.3.0](https://github.com/TryQuiet/quiet/compare/@quiet/common@1.2.1-alpha.5...@quiet/common@1.3.0) (2023-06-02)
+
+**Note:** Version bump only for package @quiet/common
+
+
+
+
+
+## [1.2.1-alpha.5](https://github.com/TryQuiet/quiet/compare/@quiet/common@1.2.1-alpha.4...@quiet/common@1.2.1-alpha.5) (2023-06-02)
+
+**Note:** Version bump only for package @quiet/common
+
+
+
+
+
 ## [1.2.1-alpha.4](https://github.com/TryQuiet/quiet/compare/@quiet/common@1.2.1-alpha.3...@quiet/common@1.2.1-alpha.4) (2023-05-25)
 
 **Note:** Version bump only for package @quiet/common